debugged the code for better yield
It was a silly mistake again, i exchanged height and width again. I have made this mistake so many times that i feel dumb at this point. Also, manually counted the yield for the old code and the newer code, the old code gave 28% yield but the new one gave 41%. Although this is by no means an accurate calculation over the big dataset, it does increase the yield a bit, most probably the far off faces.