refactored bbox
I have maybe corrected the dlib.rectangle bbox. Or messed it up. I flipped it.   I made ymin ("top") the y value for top left point as top. It executes. Going to run it on 100K files, and try to assemble video from it to see if it is working.