makevideo is working!
I was sorting twice. I was sorting the results of the test (the distance of enc1 to all possible enc2 for that round). Appending that to the face_distances list. Then after turning the face_distances list into a df, I sorted that. So it sorted on the delta(?) distance -- the distance of the file to the file before it in the correctly sorted order. SO the older white woman with grey hair (1189143918) was very very closely matched to the white man (1189143918) that precedes her in the correct sort order. And as a result, she went to he top of the list, after the start image (which has a distance of 0, obviously)