Added Object detection and Image inpainting
Object detection is done using "calc_phone_bbox" which calculated the phone_bbox, confidence and feeds it into the table "PhoneBbox" updated in "my_declarative_base" . ALso added image inpainting while runtime which check if the image has been inpainted before and if not inpaints runtime and outputs. A folder "inpaint" will be created in the root directory which will have the inpainted images. ALso you had a few combinations of image_edge_multipliers in "make_video" for "just face", "bigger portrait" etc. i took the maximum for them in all directions to inpaint the image so that any of those image_edge_multipliers can be used without changing anything in the inpainting algorithm. Damnnn this was a lot LOL