handles E11000 duplicate key errors
this is where mongo has the wrong encoding_id. if the passed eid is None, this fix checks the eids in both mongo and mysql. if they are not the same, it replaces mongo value with mysql value, and then continues with the the correct mysql value for the upsert.